requirements.txt
fastmcp>=2.0.0
mysql-connector-python>=8.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
